+++ src/framework/helpers/internalServerError.ts
@@ -1,4 +1,0 @@
-let { ApolloError } = require("apollo-server")
-export default function (e: any, code = 500) {
-  return new ApolloError(e.message, code, {})
-}