+++ src/framework/storage/generateMulterInstancesForStorages.ts
@@ -1,5 +1,0 @@
-import { IConfiguration } from "src/framework/types/configuration"
-
-export default async (configuration: IConfiguration) => {
-  let storages = configuration.storage.storageDirectory
-}