+++ src/next/devServers/dev-naveed.ts
@@ -1,9 +1,0 @@
-import wertik from "../index"
-
-const devNaveed = async () => {
-  wertik({
-    port: 1200,
-  })
-}
-
-export default devNaveed