+++ src/types/mailer.ts
@@ -1,0 +1,12 @@
+export interface SendEmailProps {
+  template: string
+  variables: {
+    [key: string]: any
+  }
+  from: string
+  to: string
+  subject: string
+  [key: string]: any
+}
+
+export interface userMailerProps {}